Conditionally require pino if logger is enabled
