+++ lib/fourOhFour.js
@@ -18,7 +18,7 @@ const { buildErrorHandler } = require('./error-handler.js')
 const {
   FST_ERR_NOT_FOUND
 } = require('./errors')
-const { createChildLogger } = require('./logger')
+const { createChildLogger } = require('./logger-utils.js')
 const { getGenReqId } = require('./reqIdGenFactory.js')
 
 /**
