+++ lib/logger.js
@@ -6,7 +6,6 @@
  * License: MIT (https://raw.githubusercontent.com/pinojs/pino-http/master/LICENSE)
  */
 
-const nullLogger = require('abstract-logging')
 const pino = require('pino')
 const { serializersSym } = pino.symbols
 const {
@@ -73,13 +72,6 @@ function now () {
 }
 
 function createLogger (options) {
-  // if no logger is provided, then create a default logger
-  if (!options.loggerInstance && !options.logger) {
-    const logger = nullLogger
-    logger.child = () => logger
-    return { logger, hasLogger: false }
-  }
-
   if (options.logger && options.loggerInstance) {
     throw new FST_ERR_LOG_LOGGER_AND_LOGGER_INSTANCE_PROVIDED()
   }
@@ -147,42 +139,9 @@ function validateLogger (logger, strict) {
   }
 }
 
-/**
- * Utility for creating a child logger with the appropriate bindings, logger factory
- * and validation.
- * @param {object} context
- * @param {import('../fastify').FastifyBaseLogger} logger
- * @param {import('../fastify').RawRequestDefaultExpression<any>} req
- * @param {string} reqId
- * @param {import('../types/logger.js').ChildLoggerOptions?} loggerOpts
- */
-function createChildLogger (context, logger, req, reqId, loggerOpts) {
-  const loggerBindings = {
-    [context.requestIdLogLabel]: reqId
-  }
-  const child = context.childLoggerFactory.call(context.server, logger, loggerBindings, loggerOpts || {}, req)
-
-  // Optimization: bypass validation if the factory is our own default factory
-  if (context.childLoggerFactory !== defaultChildLoggerFactory) {
-    validateLogger(child, true) // throw if the child is not a valid logger
-  }
-
-  return child
-}
-
-/**
- * @param {import('../fastify.js').FastifyBaseLogger} logger
- * @param {import('../types/logger.js').Bindings} bindings
- * @param {import('../types/logger.js').ChildLoggerOptions} opts
- */
-function defaultChildLoggerFactory (logger, bindings, opts) {
-  return logger.child(bindings, opts)
-}
-
 module.exports = {
   createLogger,
-  createChildLogger,
-  defaultChildLoggerFactory,
   serializers,
-  now
+  now,
+  validateLogger,
 }
