+++ lib/route.js
@@ -49,7 +49,7 @@ const {
   kRouteContext
 } = require('./symbols.js')
 const { buildErrorHandler } = require('./error-handler')
-const { createChildLogger } = require('./logger')
+const { createChildLogger } = require('./logger-utils.js')
 const { getGenReqId } = require('./reqIdGenFactory.js')
 
 function buildRouting (options) {
